requirements.txt
django>=2
djangorestframework>=3.7.7
markdown>=2.6.7
django-filter>=0.15.3
django-nose
